src/main/webapp/app/entities/BulutAttendance/account-hesab/delete/account-hesab-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { AccountHesabService } from '../service/account-hesab.service';

import { AccountHesabDeleteDialogComponent } from './account-hesab-delete-dialog.component';

describe('AccountHesab Management Delete Component', () => {
  let comp: AccountHesabDeleteDialogComponent;
  let fixture: ComponentFixture<AccountHesabDeleteDialogComponent>;
  let service: AccountHesabService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, AccountHesabDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(AccountHesabDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(AccountHesabDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(AccountHesabService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('9fec3727-3421-4967-b213-ba36557ca194');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('9fec3727-3421-4967-b213-ba36557ca194');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
